comentado dbset repetido

+++ KIS.System.Advanced.Domain/Entities/Pedido .cs
@@ -11,7 +11,7 @@ namespace KIS.System.Advanced.Domain.Entities
     {
         public int ID_PEDIDO { get; set; }
         public int ID_USUARIO_PEDIDO { get; set; }
-        public float TOTAL_PEDIDO { get; set; }
+        public Double TOTAL_PEDIDO { get; set; }
         public string OBS_PEDIDO { get; set; }
         public int DATA_REG_PEDIDO { get; set; }
         public int ID_CLIENTE { get; set; }
